Finished adding user specific media to MediaService

+++ src/main/java/com/BryceJensenius/MediaOrganizer/service/MediaServiceImplementation.java
@@ -1,13 +1,14 @@
 package com.BryceJensenius.MediaOrganizer.service;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.BryceJensenius.MediaOrganizer.model.MediaItem;
+import com.BryceJensenius.MediaOrganizer.model.User;
 import com.BryceJensenius.MediaOrganizer.repository.MediaRepository;
+import com.BryceJensenius.MediaOrganizer.repository.UserRepository;
 
 @Service
 public class MediaServiceImplementation implements MediaService {
@@ -15,23 +16,37 @@ public class MediaServiceImplementation implements MediaService {
     @Autowired
     private MediaRepository mediaRepository;
 
+    @Autowired
+    private UserRepository userRepository;
+
     @Override
-    public MediaItem saveMedia(MediaItem media) {
-        return mediaRepository.save(media);
+    public MediaItem saveMedia(MediaItem media, User user) {
+        user.addMedia(media);
+        media = mediaRepository.save(media);
+        userRepository.save(user);
+        return media;
     }
 
     @Override
-    public List<MediaItem> getAllMedia() {
-        return mediaRepository.findAll();
+    public List<MediaItem> getAllMedia(User user) {
+        return user.getMediaList();
     }
 
     @Override
-    public Optional<MediaItem> getMediaById(int id) {
-        return mediaRepository.findById(id);
+    public MediaItem getMediaById(int id, User user) {
+        MediaItem item = mediaRepository.findById(id);
+        if(item == null || item.getUser() == null || item.getUser().getId() != user.getId()){
+            return null;
+        }
+        return item; // Only return if the media item belongs to the user
     }
 
     @Override
-    public void deleteById(int id){
-        mediaRepository.deleteById(id);
+    public void deleteById(int id, User user){
+        MediaItem item = mediaRepository.findById(id);
+        if(item == null || item.getUser() == null || item.getUser().getId() != user.getId()){
+            return;
+        }
+        mediaRepository.deleteById(id); // Only delete if the media item belongs to the user
     }
 }
